Fix some integration test scaffolding.

+++ Justfile
@@ -1,4 +1,4 @@
-fixtures_url := "https://github.com/Fingel/FeFits/releases/download/test-fixtures/fixtures.tar.xz"
+fixtures_url := "https://github.com/Fingel/FeFits/releases/download/integration-tests/fixtures.tar.xz"
 fixtures_archive := "tests/fixtures.tar.xz"
 fixtures_dir := "tests/fixtures"
 
@@ -11,7 +11,7 @@ test-integration: fetch-test-data
 # Download and extract integration test fixtures
 fetch-test-data:
     mkdir -p {{ fixtures_dir }}
-    curl -L --fail -o {{ fixtures_archive }} {{ fixtures_url }}
+    [ -f {{ fixtures_archive }} ] || curl -L --fail -o {{ fixtures_archive }} {{ fixtures_url }}
     tar -xJf {{ fixtures_archive }} -C {{ fixtures_dir }}
 
 # Remove downloaded test fixtures
